% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_comparison_functions.R
\name{Add_RI_data}
\alias{Add_RI_data}
\title{Adds database RI values to input dataframe}
\usage{
Add_RI_data(Full_data, RI_data)
}
\arguments{
\item{Full_data}{A dataframe with at least one column of CAS numbers with column label "CAS", and one column of calculated RI values with column label "RI", one column of compound names called "Name",  and one column of sample identifiers called "File".}

\item{RI_data}{The output of `nist_ri` or `Get_kovats_RIs`}
}
\description{
This function takes a raw dataframe of GC output and RI database search results from that dataframe and combines them into a new dataframe no loss of information.
}
\examples{
Add_RI_data(Sample_data, Sample_RI)
}
