% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_comparison_functions.R
\name{Get_kovats_RIs}
\alias{Get_kovats_RIs}
\title{Get Kovats RIs from NIST database}
\usage{
Get_kovats_RIs(Full_data, type = "kovats", polarity = "non-polar")
}
\arguments{
\item{Full_data}{A dataframe with at least one column of CAS numbers, with column label "CAS".}

\item{type}{The type of RI to be collected from NIST. Defaults to "kovats". Other options are "linear", "alkane", or "lee".}

\item{polarity}{The polarity of the GC column. Defaults to "non-polar". Other option is "polar".}
}
\description{
This is a wrapper function for the webchem function `nist_ri()`, which scrapes the NIST database for all RI data for all compounds in a given input file. It is not as efficient as it could be, because the nist_ri function needs to be called 3 times, once for each temperature program option. For some reason, it does not take a vector as an argument (using a vector of parameters doesn't throw an error, but only the first parameter is used). Will take some time to run, particularly if the number of unique compounds in the dataset is large.
}
\examples{
Get_kovats_RIs(Full_data = na.omit(Sample_data))
}
