% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_comparison_functions.R
\name{Match_RI_data}
\alias{Match_RI_data}
\title{Makes dataframe of good RI database matches}
\usage{
Match_RI_data(Full_data, RI_data, Match_threshold = 50)
}
\arguments{
\item{Full_data}{A dataframe with at least one column of CAS numbers with column label "CAS", and one column of calculated RI values with column label "RI", one column of compound names called "Name",  and one column of sample identifiers called "File".}

\item{RI_data}{The output of `nist_ri` or `Get_kovats_RIs`}

\item{Match_threshold}{The threshold at which a database RI value is considered matched with the calculated value. Defaults to 50.}
}
\description{
This function takes a raw dataframe of GC output and RI database search results from that dataframe and combines them into a new dataframe of well-matched compounds with no loss of columns.
}
\examples{
Match_RI_data(Sample_data, Sample_RI)
}
