% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_comparison_functions.R
\name{Only_database_matches}
\alias{Only_database_matches}
\title{Only compounds with database matches}
\usage{
Only_database_matches(Full_data, RI_data, Database_RI = "RI.mean")
}
\arguments{
\item{Full_data}{A dataframe with at least one column of CAS numbers with column label "CAS", and one column of calculated RI values with column label "RI", one column of compound names called "Name",  and one column of sample identifiers called "File".}

\item{RI_data}{The output of `nist_ri` or `Get_kovats_RIs`}

\item{Database_RI}{Specify the name of the column of database RI values to be compared. Defaults to "RI.mean" in line with `Summarise_RI_info`.}
}
\description{
This function is a subset of `Matched_RI_data`, but returns a dataframe of only the compounds with database matches.
}
