% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_comparison_functions.R
\name{Only_matches}
\alias{Only_matches}
\title{Only compounds with database matches}
\usage{
Only_matches(Full_data, RI_data)
}
\arguments{
\item{Full_data}{A dataframe with at least one column of CAS numbers with column label "CAS", and one column of calculated RI values with column label "RI", one column of compound names called "Name",  and one column of sample identifiers called "File".}

\item{RI_data}{The output of `nist_ri` or `Get_kovats_RIs`}
}
\description{
This function is a subset of `Add_RI_data`, but returns a dataframe of only the compounds with database information available, regardless of whether or not the match is good.
}
\examples{
Only_matches(Sample_data, Sample_RI)
}
