% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_comparison_functions.R
\name{Poor_or_no_matches}
\alias{Poor_or_no_matches}
\title{All compounds with no or poor database matches}
\usage{
Poor_or_no_matches(Full_data, RI_data, Threshold = 50)
}
\arguments{
\item{Full_data}{A dataframe with at least one column of CAS numbers with column label "CAS", and one column of calculated RI values with column label "RI", one column of compound names called "Name",  and one column of sample identifiers called "File".}

\item{RI_data}{The output of `nist_ri` or `Get_kovats_RIs`}

\item{Threshold}{The threshold above which an RI match is considered poor. Defaults to 50.}
}
\description{
This function is simply the inverse of `Match_RI_data`, and returns a dataframe of all compounds which are missing or are matched poorly.
}
\examples{
Poor_or_no_matches(Sample_data, Sample_RI)
}
