% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_comparison_functions.R
\name{Summarise_RI_info}
\alias{Summarise_RI_info}
\title{Summarises the output of a NIST database search}
\usage{
Summarise_RI_info(RI_data)
}
\arguments{
\item{RI_data}{The output of `nist_ri` or `Get_kovats_RIs`}
}
\description{
This function takes output of `nist_ri` or `Get_kovats_RIs` as an input and returns a matrix of CAS number, mean, standard deviation of the mean, and number of database samples.
}
\examples{
Summarise_RI_info(Sample_RI)
}
